# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/schedule": {
      "post": {
        "description": "Creating a new schedule",
        "operationId": "ScheduleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Schedule has been successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateScheduleDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule is not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Schedule"
        ]
      },
      "get": {
        "description": "Get all schedule",
        "operationId": "ScheduleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of schedule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule is not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/api/schedule/{id}": {
      "get": {
        "description": "Get schedule by id",
        "operationId": "ScheduleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule doesn`t exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Schedule"
        ]
      },
      "patch": {
        "description": "Updating contact by id",
        "operationId": "ScheduleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schedule has been successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule doesn`t exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Schedule"
        ]
      },
      "delete": {
        "description": "Deleting user by id",
        "operationId": "ScheduleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Schedule has been deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Contact doesn`t exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/api/schedule/user/{id}": {
      "get": {
        "description": "Get schedule by user's id",
        "operationId": "ScheduleController_getUserSchedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule/User doesn`t exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/api/groups": {
      "post": {
        "description": "Creating a new group",
        "operationId": "GroupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group has been successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateGroupDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Groups are not found"
          },
          "409": {
            "description": "Group with given email already exists"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Groups"
        ]
      },
      "get": {
        "description": "Get all groups",
        "operationId": "GroupsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of groups",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Groups are not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Groups"
        ]
      }
    },
    "/api/groups/{id}": {
      "get": {
        "description": "Get group by id",
        "operationId": "GroupsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Group doesn`t exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Groups"
        ]
      },
      "patch": {
        "description": "Updating group by id",
        "operationId": "GroupsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group has been successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Group doesn`t exist"
          },
          "409": {
            "description": "Group with given name already exists"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Groups"
        ]
      },
      "delete": {
        "description": "Deleting group by id",
        "operationId": "GroupsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Group has been deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Group doesn`t exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "",
        "tags": [
          "Groups"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vocal Studio",
    "description": "The Vocal Studio API description",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "permanent"
          },
          "date": {
            "type": "string",
            "example": "16-11"
          },
          "time": {
            "type": "string",
            "example": "15:00"
          },
          "place": {
            "type": "string",
            "example": "place name"
          },
          "durationMin": {
            "type": "string",
            "example": "90"
          },
          "activity": {
            "type": "string",
            "example": "vocal"
          },
          "groups": {
            "type": "string",
            "example": [
              1,
              2,
              3
            ]
          }
        },
        "required": [
          "type",
          "date",
          "time",
          "place",
          "durationMin",
          "activity",
          "groups"
        ]
      },
      "ScheduleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "type": {
            "type": "string",
            "example": "permanent"
          },
          "date": {
            "type": "string",
            "example": "16-11"
          },
          "time": {
            "type": "string",
            "example": "15:00"
          },
          "place": {
            "type": "string",
            "example": "place name"
          },
          "durationMin": {
            "type": "string",
            "example": "90"
          },
          "activity": {
            "type": "string",
            "example": "vocal"
          },
          "groups": {
            "type": "string",
            "example": [
              1,
              2,
              3
            ]
          }
        },
        "required": [
          "type",
          "date",
          "time",
          "place",
          "durationMin",
          "activity",
          "groups"
        ]
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "permanent"
          },
          "date": {
            "type": "string",
            "example": "16-11"
          },
          "time": {
            "type": "string",
            "example": "15:00"
          },
          "place": {
            "type": "string",
            "example": "place name"
          },
          "durationMin": {
            "type": "string",
            "example": "90"
          },
          "activity": {
            "type": "string",
            "example": "vocal"
          },
          "groups": {
            "type": "string",
            "example": [
              1,
              2,
              3
            ]
          }
        }
      },
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Crystal"
          },
          "users": {
            "type": "string",
            "example": [
              1,
              2
            ]
          },
          "schedules": {
            "type": "string",
            "example": [
              1,
              2
            ]
          }
        },
        "required": [
          "name",
          "users",
          "schedules"
        ]
      },
      "GroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Crystal"
          },
          "users": {
            "type": "string",
            "example": [
              1,
              2
            ]
          },
          "schedules": {
            "type": "string",
            "example": [
              1,
              2
            ]
          }
        },
        "required": [
          "name",
          "users",
          "schedules"
        ]
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Crystal"
          },
          "users": {
            "type": "string",
            "example": [
              1,
              2
            ]
          },
          "schedules": {
            "type": "string",
            "example": [
              1,
              2
            ]
          }
        }
      },
...
    }
  }
}